app/blog_service/requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
